rename states/ -> scenes/ & clean up index.js
