+++ src/index.js
@@ -1,17 +1,16 @@
 import Phaser from 'phaser'
 import config from './config'
-import { Preload, MainMenu, Playing } from './states'
+import { Preload, MainMenu, Playing } from './scenes'
 
 import 'Assets/style/style.css'
 
-var gameConfig = {
+// eslint-disable-next-line no-new
+new Phaser.Game({
   type: Phaser.AUTO,
   width: config.width,
   height: config.height,
   parent: 'gameContainer',
   scenes: [Preload, MainMenu, Playing]
-}
-
-new Phaser.Game(gameConfig) // eslint-disable-line no-new
+})
 
 require('./music')
